单独一个类放在一个文件
