+++ src/dotnetCampus.Ipc/IpcRequestParameter.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace dotnetCampus.Ipc
+{
+    public class IpcRequestParameter
+    {
+        public Type ParameterType { set; get; }
+
+        public object Value { set; get; }
+    }
+}
